backend/requirements.txt
# SmartShop AI Backend - Requirements
# Python 3.10+ required

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# YOLOv8 Object Detection
ultralytics>=8.0.0

# Image Processing (use flexible versions for Windows compatibility)
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Google Gemini AI (for Vision OCR)
google-generativeai>=0.3.0

# Text-to-Speech
gTTS>=2.3.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Logging
loguru>=0.7.0
